csp_thread: Rename csp_thread.h to csp_sleep.h
csp_thread.h only has csp_sleep_ms().  Rename the file to the one
which describes what it has.
