+++ include/csp/interfaces/csp_if_udp.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <csp/csp.h>
-#include <csp/arch/csp_thread.h>
+#include <csp/arch/csp_sleep.h>
 
 #include <pthread.h>
 #include <netinet/in.h>
