+++ src/drivers/can/can_socketcan.c
@@ -15,7 +15,7 @@
 #endif
 
 #include <csp/csp.h>
-#include <csp/arch/csp_thread.h>
+#include <csp/arch/csp_sleep.h>
 
 // CAN interface data, state, etc.
 typedef struct {
